ghal: Fix quote tagging in parsed sentences
It seems that the "prose" library we're depending on isn't correctly
tokenizing quote symbols, instead marking them with all sorts of weird
tags.

This new fixup helper uses quote counting to try to assign the tags that
the "prose" docs claim it produces, while also recognizing the documented
behavior so we can avoid messing it up if the upstream behavior improves
in a future release.